requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
gunicorn==21.2.0